frontend/requirements.txt
streamlit>=1.20.0
requests>=2.28.0
plotly>=5.10.0 # Keep if you have other plots, otherwise optional
pandas>=1.5.0
python-dotenv>=1.0.0
# --- Geospatial Libraries ---
geopandas>=0.12.0 # Or latest version
folium>=0.14.0    # Or latest version
streamlit-folium>=0.10.0 # Or latest version
# Pillow is often a dependency of geopandas/folium, but good to list
Pillow>=9.0.0